Move nav into header and use header directly in views
