+++ client/reader/user-stream/components/user-profile-header.tsx
@@ -1,3 +1,5 @@
+import page from '@automattic/calypso-router';
+import { useTranslate } from 'i18n-calypso';
 import ReaderAuthorLink from 'calypso/blocks/reader-author-link';
 import ReaderAvatar from 'calypso/blocks/reader-avatar';
 import SectionNav from 'calypso/components/section-nav';
@@ -5,49 +7,58 @@ import NavItem from 'calypso/components/section-nav/item';
 import NavTabs from 'calypso/components/section-nav/tabs';
 import { UserData } from 'calypso/lib/user/user';
 
-interface NavigationItem {
-	label: string;
-	path: string;
-	selected: boolean;
-}
-
 interface UserProfileHeaderProps {
 	user: UserData;
-	navigationItems: NavigationItem[];
-	selectedTab: string;
 }
 
-const UserProfileHeader = ( {
-	user,
-	navigationItems,
-	selectedTab,
-}: UserProfileHeaderProps ): JSX.Element => (
-	<>
-		<header className="user-profile__header">
-			<ReaderAvatar author={ { ...user, has_avatar: !! user.avatar_URL } } />
-			<div className="user-profile-header__details">
-				<div className="user-profile-header__display-name">
-					<ReaderAuthorLink author={ { name: user.display_name } }>
-						{ user.display_name }
-					</ReaderAuthorLink>
-				</div>
-				{ user.bio && (
-					<div className="user-profile-header__bio">
-						<p>{ user.bio }</p>
+const UserProfileHeader = ( { user }: UserProfileHeaderProps ): JSX.Element => {
+	const translate = useTranslate();
+	const currentPath = page.current;
+	const userId = user.ID;
+
+	const navigationItems = [
+		{
+			label: translate( 'Posts' ),
+			path: `/read/users/${ userId }`,
+			selected: currentPath === `/read/users/${ userId }`,
+		},
+		{
+			label: translate( 'Lists' ),
+			path: `/read/users/${ userId }/lists`,
+			selected: currentPath === `/read/users/${ userId }/lists`,
+		},
+	];
+
+	const selectedTab = navigationItems.find( ( item ) => item.selected )?.label || '';
+
+	return (
+		<>
+			<header className="user-profile__header">
+				<ReaderAvatar author={ { ...user, has_avatar: !! user.avatar_URL } } />
+				<div className="user-profile-header__details">
+					<div className="user-profile-header__display-name">
+						<ReaderAuthorLink author={ { name: user.display_name } }>
+							{ user.display_name }
+						</ReaderAuthorLink>
 					</div>
-				) }
-			</div>
-		</header>
-		<SectionNav selectedText={ selectedTab }>
-			<NavTabs>
-				{ navigationItems.map( ( item ) => (
-					<NavItem key={ item.path } path={ item.path } selected={ item.selected }>
-						{ item.label }
-					</NavItem>
-				) ) }
-			</NavTabs>
-		</SectionNav>
-	</>
-);
+					{ user.bio && (
+						<div className="user-profile-header__bio">
+							<p>{ user.bio }</p>
+						</div>
+					) }
+				</div>
+			</header>
+			<SectionNav selectedText={ selectedTab }>
+				<NavTabs>
+					{ navigationItems.map( ( item ) => (
+						<NavItem key={ item.path } path={ item.path } selected={ item.selected }>
+							{ item.label }
+						</NavItem>
+					) ) }
+				</NavTabs>
+			</SectionNav>
+		</>
+	);
+};
 
 export default UserProfileHeader;
