+++ client/reader/user-stream/views/lists.tsx
@@ -1,17 +1,19 @@
 import { formatListBullets, Icon } from '@wordpress/icons';
 import { useTranslate } from 'i18n-calypso';
 import EmptyContent from 'calypso/components/empty-content';
+import { UserData } from 'calypso/lib/user/user';
+import UserProfileHeader from 'calypso/reader/user-stream/components/user-profile-header';
 
 interface UserListsProps {
-	headerContent: React.ReactNode;
+	user: UserData;
 }
 
-const UserLists = ( { headerContent }: UserListsProps ): JSX.Element => {
+const UserLists = ( { user }: UserListsProps ): JSX.Element => {
 	const translate = useTranslate();
 
 	return (
 		<div className="user-stream__lists">
-			{ headerContent }
+			<UserProfileHeader user={ user } />
 			<EmptyContent
 				illustration={ null }
 				icon={ <Icon icon={ formatListBullets } size={ 48 } /> }
